refactor: collectionView, ViewController 에서 분리 뷰컨트롤러에서 하는일을 줄이기 위해 collectionView 기능 class 로 분리
