+++ OpenMarket/OpenMarket/CollectionView/OpenMarketDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  OpenMarketDelegate.swift
+//  OpenMarket
+//
+//  Created by tae hoon park on 2021/09/07.
+//
+
+import UIKit
+
+class OpenMarketDelegate: NSObject, UICollectionViewDelegate {
+    
+}
