+++ OpenMarket/OpenMarket/Controller/OpenMarketViewController.swift
@@ -8,59 +8,18 @@ import UIKit
 
 class OpenMarketViewController: UIViewController {
     @IBOutlet weak var collectionView: UICollectionView!
-    
-    private var productList: [Product] = []
-    private let networkManager = NetworkManager()
-    private let parsingManager = ParsingManager()
+    let openMarketDataSource = OpenMarketDataSource()
+   
     private let compositionalLayout = CompositionalLayout()
-    private let page = 1
     
     override func viewDidLoad() {
         super.viewDidLoad()
      
-        collectionView.dataSource = self
-        collectionView.delegate = self
+        collectionView.dataSource = openMarketDataSource
+//        collectionView.delegate = self
         collectionView.register(UINib(nibName: ProductCell.listNibName, bundle: nil), forCellWithReuseIdentifier: ProductCell.identifier)
-        collectionView.register(UINib(nibName: ProductCell.GridNibName, bundle: nil), forCellWithReuseIdentifier: ProductCell.identifier)
+//        collectionView.register(UINib(nibName: ProductCell.GridNibName, bundle: nil), forCellWithReuseIdentifier: ProductCell.identifier)
         collectionView.collectionViewLayout = compositionalLayout.creat(horizontalNumber: 1, verticalSize: 100, scrollDirection: .vertical)
-        requestProductList()
-    }
-}
-
-extension OpenMarketViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        productList.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: ProductCell.identifier, for: indexPath) as? ProductCell else {
-            return  UICollectionViewCell()
-        }
-        let productForItem = productList[indexPath.item]
-        cell.imageConfigure(product: productForItem)
-        cell.textConfigure(product: productForItem)
-        
-        return cell
-    }
-}
-
-extension OpenMarketViewController: UICollectionViewDelegate {
-    
-}
-
-extension OpenMarketViewController {
-    func requestProductList() {
-        self.networkManager.commuteWithAPI(api: GetItemsAPI(page: page)) { result in
-            if case .success(let data) = result {
-                guard let product = try? self.parsingManager.decodedJSONData(type: ProductCollection.self, data: data) else {
-                    return
-                }
-                self.productList.append(contentsOf: product.items)
-                
-                DispatchQueue.main.async {
-                    self.collectionView.reloadData()
-                }
-            }
-        }
+        openMarketDataSource.requestProductList(collectionView: collectionView)
     }
 }
